♻️ refactor : 폴더 재구조화 - adapter.out.kafka.event
