--- src/main/java/com/zipte/notifications/server/adapter/in/consumer/dto/EventType.java
@@ -1,6 +1,0 @@
-package com.zipte.notifications.server.adapter.in.consumer.dto;
-
-public enum EventType {
-    ADD,
-    REMOVE
-}
